requirements.txt
alembic==1.4.2
appdirs==1.4.4
arrow==0.17.0
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.5
blessed==1.17.11
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
Django==3.1.5
django-paypal==1.0.0
django-picklefield==3.0.1
django-q==1.3.4
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django-toggle-switch-widget==0.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dnspython==2.0.0
easy-thumbnails==2.7
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.2.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PyJWT==2.0.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.5
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.3
virtualenv==20.0.28
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
